Deleted content layer on SomaticHealth component
